programas/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pandas
numpy
xarray
netCDF4
requests
earthaccess
# opcional:
metpy
